requirements.txt
asttokens==2.4.1
autopep8==2.3.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.1.1
executing==2.1.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.0
imbalanced-learn==0.13.0
imblearn==0.0
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
Pygments==2.18.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
setuptools==80.8.0
six==1.17.0
sklearn-compat==0.1.3
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenize_rt==6.1.0
tomli==2.1.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
